requirements.txt
hydra-core==1.3.2
matplotlib==3.5.1
numpy==1.25.1
omegaconf==2.3.0
pandas==1.3.5
scikit_learn==1.3.0
# torch==2.0.1+cu118
# torchvision==0.15.2+cu118
tqdm==4.64.1
uproot==5.0.10
# seaborn==0.12.2
plotly==5.15.0
tenacity==8.2.2
ipywidgets==8.1.0
ipympl==0.9.3
mpl_interactions==0.23.1
wandb==0.15.5
scipy==1.8.0
